test(backend): move resend tests to own file — zero shared surface with org-UI stack
routes_test.py is heavily modified by the in-flight org-UI stack; appending
there made the rollup eject this PR on a test-file conflict.
